Simplify code and reduce duplication
